Found reason why Lyra give me a runtime error
The Lyra instance is sent from the server to client as a JSON file. But some data from the Lyra instance are lost during the JSON conversion (e.g methods). To fix this, I'll have to create and insert data to the Lyra instance on the client. This means I have to send the whole Lyra bundle to the client, which is going to be huge. Let's try exploring other alternative